Started loading the base library automatically. Ever since I moved all the precedence declarations into base, it has been impossible to do anything even moderately nontrivial without it loaded.
In order to make this work accross different directories, I added support for a TPL_PATH environment variable that tells TPL where to find the libraries. In the future, I want cabal install to set this variable correctly when installed, but for now it will have to be set manually.